requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
pillow>=10.0.0
requests>=2.31.0

# Production dependencies
gunicorn>=21.2.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
